% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{active_set}
\alias{active_set}
\title{Change the active set of a MALECOT project}
\usage{
active_set(project, set)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{set}{the new active set}
}
\description{
Change the active set of a MALECOT project
}
\examples{
# TODO
}
