% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_data}
\alias{bind_data}
\title{Bind data to project}
\usage{
bind_data(proj, data, data_format = NULL, name = NULL, missing_data = -1,
  check_delete_output = TRUE)
}
\arguments{
\item{proj}{TODO}

\item{data}{TODO}

\item{data_format}{TODO}

\item{name}{TODO}

\item{missing_data}{TODO}

\item{check_delete_output}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\examples{
# TODO
}
