% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_data_biallelic}
\alias{bind_data_biallelic}
\title{Bind bi-allelic data to project}
\usage{
bind_data_biallelic(project, df, ID_col = 1, pop_col = NULL,
  data_cols = NULL, ID = NULL, pop = NULL, missing_data = -9,
  name = NULL, check_delete_output = TRUE)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{df}{a dataframe containing genetic information and optional meta-data}

\item{ID_col}{which column of the input data contains the sample IDs. If NULL
then IDs must be defined seperately through the \code{ID} argument}

\item{pop_col}{which column of the input data contains the ostensible 
population of the samples. If NULL then populations must be defined 
seperately through the \code{pop} argument}

\item{data_cols}{which columns of the input data contain genetic information.
Defaults to all remaining columns of the data once meta-data columns have 
been accounted for}

\item{ID}{sample IDs. Ignored if using the \code{ID_col} option}

\item{pop}{ostensible populations. Ignored if using the \code{pop_col} option}

\item{missing_data}{what value represents missing data. Defaults to -9. Must 
be a positive or negative integer, and cannot equal 0 or 1 as these are 
reserved for genetic data.}

\item{name}{optional name of the data set to aid in record keeping}

\item{check_delete_output}{whether to prompt the user before overwriting 
existing data}
}
\description{
Bind data in bi-allelic format to MALECOT project. Data should
  be formatted as a dataframe with samples in rows and loci in columns.
  Genetic data should be coded as 1 (homozygote REF allele), 0 (homozygote
  ALT allele), or 0.5 (heterozygote). Additional meta-data columns can be
  specified, including a column for sample IDs and a column for sampling
  population.
}
\examples{
# TODO
}
