% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_data_multiallelic}
\alias{bind_data_multiallelic}
\title{Bind multi-allelic format data to project}
\usage{
bind_data_multiallelic(project, df, pop = NULL, missing_data = -9,
  name = NULL, check_delete_output = TRUE)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{df}{a dataframe containing genetic information and optional meta-data}

\item{pop}{ostensible populations, if not using the \code{pop_col} option}

\item{missing_data}{what value represents missing data (defaults to -9). Must
be a positive or negative integer, and cannot equal 0 or 1 as these are 
reserved for genetic data.}

\item{name}{optional name of the data set to aid in record keeping}

\item{check_delete_output}{whether to prompt the user before overwriting 
existing data}
}
\description{
Bind data in multi-allelic format to MALECOT project. Data 
  should be formatted as a dataframe with three columns: "sample_ID", "locus"
  and "haplotype". Each row of this dataframe specifies a haplotype that was
  observed at that locus in that individual (i.e. long format). Haplotypes
  should be coded as positive integers.
}
\examples{
# TODO
}
