% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_MALECOT_loaded}
\alias{check_MALECOT_loaded}
\title{Check that MALECOT package has loaded successfully}
\usage{
check_MALECOT_loaded()
}
\description{
Simple function to check that MALECOT package has loaded 
  successfully. Prints "MALECOT loaded successfully!" if so.
}
