% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{delete_set}
\alias{delete_set}
\title{Delete MALECOT parameter set}
\usage{
delete_set(proj, set = NULL, check_delete_output = TRUE)
}
\arguments{
\item{proj}{the current MALECOT project}

\item{set}{numerical index of which set to delete}

\item{check_delete_output}{whether to perform check before deleting output corresponding to this parameter set}
}
\description{
TODO
}
\details{
TODO
}
\examples{
# TODO
}
