% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{example_mcmc}
\alias{example_mcmc}
\title{Example MCMC}
\usage{
example_mcmc(x, K = 3, mu_prior_mean = 0, mu_prior_var = 1000,
  sigma = 1, burnin = 100, samples = 1000, rungs = 10,
  solve_label_switching_on = TRUE, coupling_on = TRUE, scaffold_on = TRUE,
  scaffold_n = 10, splitmerge_on = TRUE, cluster = NULL)
}
\arguments{
\item{x}{the raw data (vector)}

\item{K}{the number of mixture components}

\item{mu_prior_mean}{the mean of the (normal) prior on mixture component locations}

\item{mu_prior_var}{the variance of the (normal) prior on mixture component locations}

\item{sigma}{the standard deviation of mixture components}

\item{burnin}{the number of burn-in iterations}

\item{samples}{the number of sampling iterations}

\item{rungs}{the number of temperature rungs}

\item{solve_label_switching_on}{whether to implement the Stevens' solution to the label-switching problem}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature rungs}

\item{scaffold_on}{whether to use scaffolds to improve mixing}

\item{scaffold_n}{the number of scaffolds to use}

\item{splitmerge_on}{whether to implement a split-merge proposal}

\item{cluster}{pass in a cluster environment}
}
\description{
Demonstrates good algorithmic and coding practices for a simple normal mixture model. Implements advanced MCMC methods, including Metropolis-coupling over temperature rungs, a split-merge proposal, and an additional form of proposal referred to here as a "scaffold" proposal. Uses Rcpp and the package parallel for speed and efficiency.
}
\details{
TODO
}
\examples{
# TODO
}
