% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fix_labels}
\alias{fix_labels}
\title{Match groupings using Hungarian algorithm}
\usage{
fix_labels(proj, target_group)
}
\arguments{
\item{proj}{TODO}

\item{target_group}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\examples{
# TODO
}
