% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_scaffolds}
\alias{generate_scaffolds}
\title{Generate scaffolds}
\usage{
generate_scaffolds(proj, scaffold_n = 10, K = NULL, burnin = 1000,
  auto_converge = TRUE, split_merge_on = TRUE, precision = 0.01,
  silent = FALSE, cluster = NULL)
}
\arguments{
\item{proj}{TODO}

\item{scaffold_n}{TODO}

\item{K}{the number of mixture components}

\item{burnin}{the number of burn-in iterations}

\item{auto_converge}{whether burn-in should be diagnosed automatically}

\item{split_merge_on}{whether to implement a split-merge proposal}

\item{precision}{the level of precision that allele frequencies are represented at}

\item{silent}{supresses output messages}

\item{cluster}{pass in a cluster environment}
}
\description{
TODO
}
\details{
TODO
}
\examples{
# TODO
}
