% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_ESS}
\alias{get_ESS}
\title{Get ESS}
\usage{
get_ESS(project, K = NULL)
}
\arguments{
\item{project}{a MALCOT project, as produced by the function 
\code{malecot_project()}}

\item{K}{get ESS for this value of K}
}
\description{
Returns effective sample size (ESS) of chosen model run.
}
\examples{
# TODO
}
