% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_group_order}
\alias{get_group_order}
\title{Match grouping against q-matrix}
\usage{
get_group_order(project, K, target_group)
}
\arguments{
\item{project}{a MALCOT project, as produced by the function 
\code{malecot_project()}}

\item{K}{compare against qmatrix output for this value of K}

\item{target_group}{the target group to be aligned against the qmatrix}
}
\description{
Compares qmatrix output for a chosen value of K against a 
  \code{target_group} vector. Returns the order of \code{target_group} 
  groups, such that there is the best possible alignment against the qmatrix.
  For example, if the vector returned is \code{c(2,3,1)} then the second
  group in the target vector should be matched against the first group in the
  qmatrix, followed by the third group in the target vector against the
  second group in the qmatrix, followed by the first group in the target
  vector against the third group in the qmatrix.
}
\examples{
# TODO
}
