% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.malecot_GTI_path}
\alias{is.malecot_GTI_path}
\title{Determine if object is of class malecot_GTI_path}
\usage{
is.malecot_GTI_path(x)
}
\arguments{
\item{x}{TODO}
}
\description{
Determine if object is of class malecot_GTI_path
}
\details{
TODO
}
\examples{
# TODO
}
