% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.malecot_loglike_quantiles}
\alias{is.malecot_loglike_quantiles}
\title{Determine if object is of class malecot_loglike_quantiles}
\usage{
is.malecot_loglike_quantiles(x)
}
\arguments{
\item{x}{TODO}
}
\description{
Determine if object is of class malecot_loglike_quantiles
}
\details{
TODO
}
\examples{
# TODO
}
