% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malecot_project.R
\name{is.malecot_project}
\alias{is.malecot_project}
\title{Determine if object is of class malecot_project}
\usage{
is.malecot_project(x)
}
\arguments{
\item{x}{TODO}
}
\description{
Determine if object is of class malecot_project.
}
\details{
TODO
}
\examples{
# TODO

}
