% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.malecot_q_matrix}
\alias{is.malecot_q_matrix}
\title{Determine if object is of class malecot_q_matrix}
\usage{
is.malecot_q_matrix(x)
}
\arguments{
\item{x}{TODO}
}
\description{
Determine if object is of class malecot_q_matrix
}
\details{
TODO
}
\examples{
# TODO
}
