% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malecot_project.R
\name{malecot_project}
\alias{malecot_project}
\title{Define empty malecot_project object}
\usage{
malecot_project()
}
\description{
Define empty malecot_project object
}
\details{
TODO
}
\examples{
# TODO

}
