% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.malecot_m_quantiles}
\alias{plot.malecot_m_quantiles}
\title{Default plot for class malecot_m_quantiles}
\usage{
\method{plot}{malecot_m_quantiles}(x, y, ...)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{...}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\examples{
# TODO
}
