% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.malecot_q_matrix}
\alias{plot.malecot_q_matrix}
\title{Default plot for class malecot_q_matrix}
\usage{
\method{plot}{malecot_q_matrix}(x, y = NULL, ...)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{...}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\examples{
# TODO
}
