% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_COI}
\alias{plot_COI}
\title{Plot COI 95\% credible intervals}
\usage{
plot_COI(project, K = NULL, ...)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{K}{which value of K to produce the plot for}

\item{...}{TODO}
}
\description{
Plot COI 95\% credible intervals of current active set
}
\details{
TODO
}
\examples{
# TODO
}
