% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_COI_mean}
\alias{plot_COI_mean}
\title{Plot COI_mean 95\% credible intervals}
\usage{
plot_COI_mean(project, K = NULL, ...)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{K}{TODO}

\item{...}{TODO}
}
\description{
Plot COI_mean 95\% credible intervals of current active set
}
\details{
TODO
}
\examples{
# TODO
}
