% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_COI_mean_quantiles}
\alias{plot_COI_mean_quantiles}
\title{Plot COI_mean quantiles of current active set}
\usage{
plot_COI_mean_quantiles(project, K = NULL, ...)
}
\arguments{
\item{project}{TODO}

\item{K}{TODO}

\item{...}{TODO}
}
\description{
Plot COI_mean quantiles of current active set
}
\details{
TODO
}
\examples{
# TODO
}
