% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coupling}
\alias{plot_coupling}
\title{Plot Metropolis-coupling acceptance rates}
\usage{
plot_coupling(project, K = NULL)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{K}{which value of K to plot}
}
\description{
Plot Metropolis-coupling acceptance rates
}
