% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_density}
\alias{plot_density}
\title{Produce MCMC density plot}
\usage{
plot_density(project, K = NULL, rung = NULL, col = "black")
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{mavproject()}}

\item{K}{which value of K to produce the plot for}

\item{rung}{which value of K to produce the plot for. Defaults to the cold rung}

\item{col}{colour of the trace}
}
\description{
Produce MCMC density plot of the log-likelihood
}
