% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_e}
\alias{plot_e}
\title{Plot error rate 95\% credible intervals}
\usage{
plot_e(project, K = NULL)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{K}{which value of K to plot}
}
\description{
Plot error rate 95\% credible intervals of current active set
}
\details{
TODO
}
\examples{
# TODO
}
