% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_evidence}
\alias{plot_evidence}
\title{Plot model evidence estimates over K}
\usage{
plot_evidence(proj, K = NULL, ...)
}
\arguments{
\item{proj}{TODO}

\item{K}{TODO}

\item{...}{TODO}
}
\description{
Plot model evidence estimates over K
}
\details{
TODO
}
\examples{
# TODO
}
