% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_p}
\alias{plot_p}
\title{Plot allele frequency 95\% credible intervals}
\usage{
plot_p(project, K = NULL, deme = 1, ...)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{K}{TODO}

\item{deme}{TODO}

\item{...}{TODO}
}
\description{
Plot allele frequency 95\% credible intervals of current active set
}
\details{
TODO
}
\examples{
# TODO
}
