% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_p_quantiles}
\alias{plot_p_quantiles}
\title{Plot p quantiles of current active set}
\usage{
plot_p_quantiles(project, K = NULL, deme = 1, ...)
}
\arguments{
\item{project}{TODO}

\item{K}{TODO}

\item{deme}{TODO}

\item{...}{TODO}
}
\description{
Plot p quantiles of current active set
}
\details{
TODO
}
\examples{
# TODO
}
