% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca}
\alias{plot_pca}
\title{Produce PCA plot from MALECOT data}
\usage{
plot_pca(data, type = "3D", missing_data = -1, target_group = NULL)
}
\arguments{
\item{data}{TODO}

\item{type}{TODO}

\item{missing_data}{TODO}

\item{target_group}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\examples{
# TODO
}
