% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior_K}
\alias{plot_posterior_K}
\title{Plot posterior K}
\usage{
plot_posterior_K(project)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}
}
\description{
Plot posterior K
}
