% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior_model}
\alias{plot_posterior_model}
\title{Plot posterior model}
\usage{
plot_posterior_model(project)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}
}
\description{
Plot posterior model
}
