% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_prior_p}
\alias{plot_prior_p}
\title{Plot prior on allele frequencies}
\usage{
plot_prior_p(lambda = 1, alleles = NULL)
}
\arguments{
\item{lambda}{shape parameter(s) of the Beta or Dirichlet distribution. Can
be a single scalar value, in which case the dimensionality is given by the
number of \code{alleles}, or a vector of values specifying the shape
parameter for each allele}

\item{alleles}{the dimensionality of the prior. Defaults to the length of
\code{lambda}, or to 2 of \code{lambda} is a scalar}
}
\description{
Produce plot of the prior on COI for given parameters. This
  prior is Beta in the bi-allelic case, and Dirichlet in the multi-allelic
  case.
}
