% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_q_matrix}
\alias{plot_q_matrix}
\title{Plot Q-matrix of current active set}
\usage{
plot_q_matrix(proj, y = NULL, K = NULL, ...)
}
\arguments{
\item{proj}{TODO}

\item{y}{TODO}

\item{K}{TODO}

\item{...}{TODO}
}
\description{
Plot Q-matrix of current active set
}
\details{
TODO
}
\examples{
# TODO
}
