% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_structure}
\alias{plot_structure}
\title{Posterior allocation plot}
\usage{
plot_structure(project, K = NULL, base_colours = col_hot_cold(),
  divide_ind_on = FALSE)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{K}{which value of K to plot}

\item{base_colours}{colours from which final plotting colours are taken.
These will be interpolated to produce final colours}

\item{divide_ind_on}{whether to add dividing lines between bars}
}
\description{
Produce posterior allocation plot of current active set.
}
