% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trace}
\alias{plot_trace}
\title{Produce MCMC trace plot}
\usage{
plot_trace(project, K = NULL, rung = NULL, col = "black")
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{K}{which value of K to produce the plot for}

\item{rung}{which value of K to produce the plot for. Defaults to the cold rung}

\item{col}{colour of the trace}
}
\description{
Produce MCMC trace plot of the log-likelihood at each iteration.
}
