% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.malecot_p_quantiles}
\alias{print.malecot_p_quantiles}
\title{Overload print function for malecot_p_quantiles}
\usage{
\method{print}{malecot_p_quantiles}(x, ...)
}
\arguments{
\item{x}{TODO}

\item{...}{TODO}
}
\description{
Overload print function for malecot_p_quantiles
}
\details{
TODO
}
\examples{
# TODO
}
