% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malecot_project.R
\name{print.malecot_project}
\alias{print.malecot_project}
\title{Custom print function for class malecot_project}
\usage{
\method{print}{malecot_project}(x, ...)
}
\arguments{
\item{x}{object of class \code{malecot_project}}

\item{...}{other arguments (ignored)}
}
\description{
Custom print function for class malecot_project, printing a 
  summary of the key elements (also equivalent to \code{summary(x)}). To do 
  an ordinary \code{print()} of all elements of the project, use the 
  \code{print_full()} function.
}
