% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print.malecot_q_matrix}
\alias{print.malecot_q_matrix}
\title{Overload print function for malecot_q_matrix}
\usage{
\method{print}{malecot_q_matrix}(x, ...)
}
\arguments{
\item{x}{TODO}

\item{...}{TODO}
}
\description{
Overload print function for malecot_q_matrix
}
\details{
TODO
}
\examples{
# TODO
}
