% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summary.malecot_GTI_path}
\alias{summary.malecot_GTI_path}
\title{Overload summary function for malecot_GTI_path}
\usage{
\method{summary}{malecot_GTI_path}(object, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}
}
\description{
Overload summary function for malecot_GTI_path
}
\details{
TODO
}
\examples{
# TODO
}
