% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summary.malecot_p_quantiles}
\alias{summary.malecot_p_quantiles}
\title{Overload summary function for malecot_p_quantiles}
\usage{
\method{summary}{malecot_p_quantiles}(object, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}
}
\description{
Overload summary function for malecot_p_quantiles
}
\details{
TODO
}
\examples{
# TODO
}
