% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malecot_project.R
\name{summary.malecot_project}
\alias{summary.malecot_project}
\title{Print summary for class malecot_project}
\usage{
\method{summary}{malecot_project}(object, ...)
}
\arguments{
\item{object}{object of class \code{malecot_project}}

\item{...}{other arguments (ignored)}
}
\description{
Overload summary function for class malecot_project
}
