% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summary.malecot_q_matrix}
\alias{summary.malecot_q_matrix}
\title{Overload summary function for malecot_q_matrix}
\usage{
\method{summary}{malecot_q_matrix}(object, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}
}
\description{
Overload summary function for malecot_q_matrix
}
\details{
TODO
}
\examples{
# TODO
}
