% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_data_multiallelic}
\alias{bind_data_multiallelic}
\title{Bind multi-allelic format data to project}
\usage{
bind_data_multiallelic(project, df, pop = NULL, missing_data = -9,
  alleles = NULL, name = NULL, check_delete_output = TRUE)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{df}{a dataframe with three columns, as decribed above}

\item{pop}{ostensible populations of the samples}

\item{missing_data}{what value represents missing data. Defaults to -9. Must
be a positive or negative integer}

\item{alleles}{the number of alleles at each locus. If scalar then the same 
number of alleles is assumed at all loci. If NULL then the number of
alleles is inferred directly from data as the maximum observed value per
locus}

\item{name}{optional name of the data set to aid in record keeping}

\item{check_delete_output}{whether to prompt the user before overwriting 
existing data}
}
\description{
Bind data in multi-allelic format to MALECOT project. Data 
  should be formatted as a dataframe with three columns: "sample_ID", "locus"
  and "haplotype". Each row of this dataframe specifies a haplotype that was 
  observed at that locus in that individual. Haplotypes should be coded as
  positive integers.
}
\examples{
# TODO
}
