% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{delete_set}
\alias{delete_set}
\title{Delete parameter set}
\usage{
delete_set(proj, set = NULL, check_delete_output = TRUE)
}
\arguments{
\item{set}{which set to delete. Defaults to the current active set}

\item{check_delete_output}{whether to prompt the user before deleting any 
existing output}

\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}
}
\description{
Delete a given parameter set from a MALECOT project.
}
\examples{
# TODO
}
