% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{new_set}
\alias{new_set}
\title{Create new MALECOT parameter set}
\usage{
new_set(project, name = "(no name)", lambda = 1, COI_model = "poisson",
  COI_max = 20, COI_manual = NULL, estimate_COI_mean = TRUE,
  COI_mean = 3, COI_dispersion = 2, estimate_error = FALSE, e1 = 0,
  e2 = 0, e1_max = 0.2, e2_max = 0.2)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{name}{the name of the parameter set}

\item{lambda}{shape parameter(s) governing the prior on allele frequencies. 
This prior is a Beta distribution for the bi-allelic case or a Dirichlet 
distribution for the multi-allelic case. In the b-allelic case, 
\code{lambda} can be either a single scalar value, in which case the same 
shape paremeter is applied to all loci and all alleles (i.e. a symmetric 
Beta prior), or \code{lambda} can be a list of length \code{L} of vectors
of length 2, allowing different shape parameters to be specified for each
allele at each locus individually. In the multi-allelic case, \code{lambda}
must be a single scalar value that applies to all loci and all alleles
(i.e. a symmetric Dirichlet prior)}

\item{COI_model}{the type of prior on COI. Must be one of "uniform",
"poisson", or "nb" (negative binomial)}

\item{COI_max}{the maximum COI allowed for any given sample}

\item{COI_manual}{A vector of length n (where n is the number of samples)
allowing the COI to be specified manually. Positive values indicate fixed
COIs that should not be updated as part of the MCMC, while -1 values
indicate that COIs should be estimated. Defaults to \code{rep(-1,n)},
meaning all COIs will be esimated}

\item{estimate_COI_mean}{whether the mean COI should be estimated for each 
subpopulation as part of the MCMC, otherwise the value \code{COI_mean} is 
used for all subpopulations. Defaults to \code{TRUE}. Note that mean COI
estimation is only possible under the "poisson" and "nb" models (see
\code{COI_model})}

\item{COI_mean}{single scalar value specifying the mean COI for all
subpopulations (see \code{estimate_COI_mean} above)}

\item{COI_dispersion}{must be > 1}

\item{estimate_error}{TODO}

\item{e1}{the probability of a true homozygote being incorrectly called as a
heterozygote}

\item{e2}{the probability of a true heterozygote being incorrectly called as a
homozygote}

\item{e1_max}{the maximum possible value of \code{e1}}

\item{e2_max}{the maximum possible value of \code{e2}}
}
\description{
TODO
}
\details{
TODO
}
\examples{
# TODO
}
