% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{new_set}
\alias{new_set}
\title{Create new MALECOT parameter set}
\usage{
new_set(proj, set_description = NULL, K_range = 1:3, lambda = 1,
  COI_model = "poisson", COI_max = 20, COI_dispersion = 1,
  estimate_error = FALSE, e1 = 0, e2 = 0, e1_max = 0.2, e2_max = 0.2)
}
\arguments{
\item{proj}{the current MALECOT project}

\item{set_description}{brief description of this parameter set}

\item{K_range}{TODO}

\item{lambda}{TODO}

\item{COI_model}{TODO}

\item{COI_max}{TODO}

\item{COI_dispersion}{TODO}

\item{estimate_error}{TODO}

\item{e1}{TODO}

\item{e2}{TODO}

\item{e1_max}{TODO}

\item{e2_max}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\examples{
# TODO
}
