% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{new_set}
\alias{new_set}
\title{Create new MALECOT parameter set}
\usage{
new_set(project, name = "(no name)", lambda = 1, COI_model = "poisson",
  COI_max = 20, COI_manual = NULL, estimate_COI_mean = TRUE,
  COI_mean = 3, COI_dispersion = 2, estimate_error = FALSE, e1 = 0,
  e2 = 0, e1_max = 0.2, e2_max = 0.2)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{name}{the name of the parameter set}

\item{lambda}{the shape parameter(s) of the prior on allele frequencies. This
prior is Beta in the bi-allelic case, and Dirichlet in the multi-allelic 
case. \code{lambda} can be:
\itemize{
  \item{a single scalar value, in which case the same value is used for 
  every allele and every locus (i.e. the prior is symmetric)}
  \item{a vector of values, in which case the same vector is used for every
  locus. Only works if the same number of alleles applies at every locus}
  \item{a list of vectors specifying the shape parameter separately for
  each allele of each locus. The list must of length \code{L}, and must
  contain vectors of length equal to the number of alleles at that locus}
}}

\item{COI_model}{the type of prior on COI. Must be one of "uniform",
"poisson", or "nb" (negative binomial)}

\item{COI_max}{the maximum COI allowed for any given sample}

\item{COI_manual}{A vector of length n (where n is the number of samples)
allowing the COI to be specified manually. Positive values indicate fixed
COIs that should not be updated as part of the MCMC, while -1 values
indicate that COIs should be estimated. Defaults to \code{rep(-1,n)},
meaning all COIs will be esimated}

\item{estimate_COI_mean}{whether the mean COI should be estimated for each 
subpopulation as part of the MCMC, otherwise the value \code{COI_mean} is 
used for all subpopulations. Defaults to \code{TRUE}. Note that mean COI 
estimation is only possible under the Poisson and negative binomial models
(see \code{COI_model})}

\item{COI_mean}{single scalar value specifying the mean COI for all
subpopulations (see \code{estimate_COI_mean} above)}

\item{COI_dispersion}{the ratio of the variance to the mean of the prior on
COI. Only applies under the negative binomial model. Must be >1, as a ratio
of 1 can be achieved by using the Poisson distribution}

\item{estimate_error}{whether to estimate error probabilities \code{e1} and
\code{e2}}

\item{e1}{the probability of a true homozygote being incorrectly called as a
heterozygote}

\item{e2}{the probability of a true heterozygote being incorrectly called as a
homozygote}

\item{e1_max}{the maximum possible value of \code{e1}}

\item{e2_max}{the maximum possible value of \code{e2}}
}
\description{
TODO
}
\details{
TODO
}
\examples{
# TODO
}
