% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Example MCMC}
\usage{
run_mcmc(proj, K = NULL, burnin = 100, samples = 1000, rungs = 1,
  auto_converge = TRUE, coupling_on = TRUE, scaffold_on = TRUE,
  scaffold_n = 10, split_merge_on = TRUE, solve_label_switching = TRUE,
  precision = 0.01, GTI_pow = 2, report_iteration = 100,
  flush_console = FALSE, silent = FALSE, output_format = 1,
  cluster = NULL)
}
\arguments{
\item{proj}{TODO}

\item{K}{the number of mixture components}

\item{burnin}{the number of burn-in iterations}

\item{samples}{the number of sampling iterations}

\item{rungs}{the number of temperature rungs}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature rungs}

\item{scaffold_on}{whether to use scaffolds to improve mixing}

\item{scaffold_n}{the number of scaffolds to use}

\item{cluster}{pass in a cluster environment}

\item{mu_prior_mean}{the mean of the (normal) prior on mixture component locations}

\item{mu_prior_var}{the variance of the (normal) prior on mixture component locations}

\item{sigma}{the standard deviation of mixture components}

\item{solve_label_switching_on}{whether to implement the Stevens' solution to the label-switching problem}

\item{splitmerge_on}{whether to implement a split-merge proposal}
}
\description{
Demonstrates good algorithmic and coding practices for a simple normal mixture model. Implements advanced MCMC methods, including Metropolis-coupling over temperature rungs, a split-merge proposal, and an additional form of proposal referred to here as a "scaffold" proposal. Uses Rcpp and the package parallel for speed and efficiency.
}
\details{
TODO
}
\examples{
# TODO
}
