% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Example MCMC}
\usage{
run_mcmc(proj, K = NULL, burnin = 100, samples = 1000, rungs = 1,
  auto_converge = TRUE, coupling_on = TRUE, scaffold_on = TRUE,
  scaffold_n = 10, split_merge_on = TRUE, solve_label_switching_on = TRUE,
  precision = 0.01, GTI_pow = 2, silent = FALSE, output_format = 1,
  cluster = NULL)
}
\arguments{
\item{proj}{TODO}

\item{K}{the number of mixture components}

\item{burnin}{the number of burn-in iterations}

\item{samples}{the number of sampling iterations}

\item{rungs}{the number of temperature rungs}

\item{auto_converge}{whether burn-in should be diagnosed automatically}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature rungs}

\item{scaffold_on}{whether to use scaffolds to improve mixing}

\item{scaffold_n}{the number of scaffolds to use}

\item{split_merge_on}{whether to implement a split-merge proposal}

\item{solve_label_switching_on}{whether to implement Stevens' solution to the label-switching problem}

\item{precision}{the level of precision that allele frequencies are represented at}

\item{GTI_pow}{the power used in the generalised thermodynamic integration method for estimating K}

\item{silent}{supresses output messages}

\item{output_format}{choose from a range of output formats}

\item{cluster}{pass in a cluster environment}
}
\description{
Demonstrates good algorithmic and coding practices for a simple normal mixture model. Implements advanced MCMC methods, including Metropolis-coupling over temperature rungs, a split-merge proposal, and an additional form of proposal referred to here as a "scaffold" proposal. Uses Rcpp and the package parallel for speed and efficiency.
}
\details{
TODO
}
\examples{
# TODO
}
