% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run main MCMC}
\usage{
run_mcmc(project, K = NULL, precision = 0.01, burnin = 1000,
  samples = 1000, rungs = 1, GTI_pow = 3, auto_converge = TRUE,
  converge_test = 100, solve_label_switching_on = TRUE,
  coupling_on = TRUE, cluster = NULL, pb_markdown = FALSE,
  store_acceptance = TRUE, store_raw = TRUE, silent = FALSE)
}
\arguments{
\item{project}{a MALECOT project, as produced by the function 
\code{malecot_project()}}

\item{K}{the values of K that the MCMC will explore}

\item{precision}{the level of precision at which allele frequencies are 
represented in the bi-allelic case. This allows the use of look-up tables
for the likelihood, which significantly speeds up the MCMC. Set to 0 to use
exact values (up to C++ "double" precision) rather than using look-up 
tables}

\item{burnin}{the number of burn-in iterations}

\item{samples}{the number of sampling iterations}

\item{rungs}{the number of temperature rungs}

\item{GTI_pow}{the power used in the generalised thermodynamic integration 
method. Must be greater than 1.1}

\item{auto_converge}{whether convergence should be assessed automatically 
every \code{converge_test} iterations, leading to termination of the 
burn-in phase. If \code{FALSE} then the full \code{burnin} iterations are 
used}

\item{converge_test}{test for convergence every \code{convergence_test} 
iterations if \code{auto_converge} is being used}

\item{solve_label_switching_on}{whether to implement the Stevens' solution to
the label-switching problem. If turned off then Q-matrix output will no 
longer be correct, although evidence estimates will be unaffected.}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature 
rungs}

\item{cluster}{option to pass in a cluster environment (see package 
"parallel")}

\item{pb_markdown}{whether to run progress bars in markdown mode, in which 
case they are updated once at the end to avoid large amounts of output}

\item{store_acceptance}{whether to store acceptance rates for all parameters
updated by Metropolis-Hastings. Proposal distributions are tuned adaptively
with a target acceptance rate of 23\%}

\item{store_raw}{whether to store raw MCMC output in addition to summary 
output. Setting to FALSE can considerably reduce output size in memory}

\item{silent}{whether to suppress all console output}
}
\description{
Run the main MALECOT MCMC. Model parameters are taken from the 
  current active parameter set, and MCMC parameters are passed in as 
  arguments. All output is stored within the project.
}
\examples{
# TODO
}
