% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate genetic data}
\usage{
sim_data(n = 100, L = 24, K = 3, data_format = "biallelic",
  alleles = 2, lambda = 1, COI_model = "poisson", COI_manual = NULL,
  COI_max = 20, COI_mean = 3, COI_dispersion = 1, e1 = 0, e2 = 0,
  prop_missing = 0)
}
\arguments{
\item{n}{TODO}

\item{L}{TODO}

\item{K}{TODO}

\item{data_format}{TODO}

\item{alleles}{TODO}

\item{lambda}{TODO}

\item{COI_model}{TODO}

\item{COI_manual}{TODO}

\item{COI_max}{TODO}

\item{COI_mean}{TODO}

\item{COI_dispersion}{TODO}

\item{e1}{TODO}

\item{e2}{TODO}

\item{prop_missing}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\examples{
# TODO
}
